feat: add `<TypingFeature />` w/o functionality
