+++ app/page.tsx
@@ -3,6 +3,7 @@ import BitTwiddlingFeature from './(components)/works/bit-twiddling';
 import ChessFeature from './(components)/works/chess';
 import CoolContractsFeature from './(components)/works/cool-contracts';
 import TxDotCoolFeature from './(components)/works/tx-dot-cool';
+import TypingFeature from './(components)/works/typing';
 
 import ContainerLayout from '@/components/layouts/container';
 
@@ -14,6 +15,7 @@ export default function Page() {
       <div className="grid grid-cols-2 gap-4 min-[560px]:grid-cols-4 min-[960px]:grid-cols-6">
         <ChessFeature />
         <TxDotCoolFeature />
+        <TypingFeature />
         <BitTwiddlingFeature />
         <CoolContractsFeature />
       </div>
